(fix): a demonstration's frequency is frames per second, not save_freq
RoboTwin steps physics at 1/250 s and records a frame every `save_freq` steps, so at the default
15 a demonstration runs at ~16.7 fps. `Demonstration.frequency` was being handed `save_freq`
itself — the same conflation upstream makes — which a model timing its context by it would
misread, and which set the clips' playback rate.

`robotwin.frame_rate_hz` reads the live scene's timestep, and the generator's parameter is now
named for what it is, `save_freq`.

Refs #3

+++ src/robotwin_icil/demo.py
@@ -55,7 +55,8 @@ class Demonstration:
     """One successful expert trajectory, as context for a frozen policy."""
 
     frames: tuple[Frame, ...]
-    frequency: int
+    # Frames per second of the recording: the sim rate over RoboTwin's `save_freq`.
+    frequency: float
     cameras: tuple[str, ...] = field(default=())
 
     def __post_init__(self) -> None:
